% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective population size.R
\name{eff.size.fn}
\alias{eff.size.fn}
\title{Effective population size estimation}
\usage{
eff.size.fn(
  threshold.N,
  surv.p,
  sex.ratio,
  maturity.age.F,
  maturity.age.M,
  max.age.F = NULL,
  max.age.M = NULL,
  method = c("BFGS", "Incremental"),
  round = T
)
}
\arguments{
\item{method}{"\code{BFGS}" or "\code{Incremental}", selects method for estimation of appropriate N0}

\item{round}{logical, if true, reported abundance are rounded to the nearest integer (\code{default = TRUE})}

\item{\code{threshold.N}}{numeric, adult population threshold target}

\item{\code{surv.p}}{A data frame, with 2 columns "\code{surv.F}" and "\code{surv.M}" giving for each age-sex class the corresponding survival probability}

\item{\code{sex.ratio}}{numeric, sex ratio of females in \code{N0}}

\item{\code{maturity.age.F}, \code{maturity.age.M}}{numeric, sexual maturity age for females and males}

\item{\code{max.age.F}, \code{max.age.M}}{numeric, maximum age for females and males. If not included, it is inferred from the survival probability data.}
}
\value{
a list, with \code{Ntot} (total number of adults), \code{NaF} (total number of adults females), \code{NaM} (total number of adults males), \code{Nx} (data frame of age-sex specific abundances), and \code{N0} (estimated initial population size for age-class 0). A list of input parameter value is also included.
}
\description{
Calculating population pyramid based on age-sex-specific survival probabilities and initial population size for age-class 0 based on target adult population threshold
}
\details{
\code{eff.size.fn} estimate an appropriate initial population size for age-class 0 (\code{N0}) based on a specified targeted adult population threshold.
Optimization method to assess correct N0 is controlled by the argument method.
Method "\code{BFGS}" uses a quasi-Newton approach (also known as a variable metric algorithm), specifically that published simultaneously in 1970 by Broyden, Fletcher, Goldfarb and Shanno. See ?optim for more details.
Method "\code{Incremental}" performs an exhaustive search for N0 using a complete incremental search starting at \code{N0=1}. For large N0, this approach can get really slow.
}
\examples{
surv.p=data.frame(  surv.F=c( 0.501, 0.624, 0.672, 0.696, 0.71 , 0.718, 0.876, 0.88 , 0.883,
                              0.885, 0.886, 0.887, 0.888, 0.888, 0.889, 0.889, 0.889, 0.89 ,
                              0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0.89 ,
                              0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0) ,
                    surv.M=c( 0.501, 0.624, 0.672, 0.696, 0.71 , 0.718, 0.876, 0.88 , 0.883,
                              0.885, 0.886, 0.887, 0.888, 0.888, 0.889, 0.889, 0.889, 0.89 ,
                              0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0.89 ,
                              0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0.89 , 0)
                  )

maturity.age.F=4
maturity.age.M=4
sex.ratio=0.444

threshold.N.adult=199595

start=Sys.time()
N.res=eff.size.fn(threshold.N=threshold.N.adult,surv.p,sex.ratio,maturity.age.F,maturity.age.M,method="BFGS")
end=Sys.time()
end-start    #  Time difference of 0.171541 secs

### Not Run: ###
start=Sys.time()
N.res=eff.size.fn(threshold.N=threshold.N.adult,surv.p,sex.ratio,maturity.age.F,maturity.age.M,method="Incremental")
end=Sys.time()
end-start   #  Time difference of 1.934507 mins
}
