% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective population size.R
\name{test.fn}
\alias{test.fn}
\title{Testing vector for contiguous survival data starting from the first element}
\usage{
test.fn(vec)
}
\arguments{
\item{\code{vec}}{vector of non-harvest mortality rate of length Y}
}
\value{
\code{TRUE} if the vector is composed either of non missing values, or start with a contiguous set of non missing values followed by NAs.
}
\description{
Testing vector for contiguous survival data starting from the first element
}
\examples{
test.fn(c(4,3,6,1,7,8))   # TRUE
test.fn(c(4,3,6,1,7,NA))  # TRUE
test.fn(c(4,3,6,NA,1,8))  # FALSE
test.fn(c(NA,4,3,6,1,8))  # FALSE
test.fn(c(4,3,NA,6,1,NA)) # FALSE
}
